Remove exceptions from the common path
Previously, exceptions were being used as a common case for control flow. This
was causing a large amount of time being consumed in the exception unwinder
code. The lookup routines as well as the indexing routines will now return NULL
if the name isn't found. The calling methods have been changed to look for a
NULL return value and handle it gracefully.